ref(sdk-crashes): Extract store event fixture
Move store event fixture to conftest.py, so it can be reused.
